add create playlist and add song to playlist

+++ model/song.go
@@ -1,9 +1,9 @@
 package model
 
 type Song struct {
-	Id       int64
+	Id       int
 	Name     string
 	Length   float64
 	Genre    string
-	ArtistId int64
+	ArtistId int
 }
